requirements.txt
tabulate==0.9.0
langchain-experimental==0.3.4
langchain-openai==0.3.2
langchain-teddynote==0.3.42
python-dotenv==1.0.1
streamlit==1.41.1
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
plotly==5.24.1
scikit-learn==1.6.1